Moved ingesters to an ingester folder
