+++ src/gluon/ingesters/sharedIngester.ts
